trip_planner/requirements.txt
fastapi
uvicorn
google-generativeai
google-adk
python-dotenv
google-auth
requests
pydantic
